// Backend/src/modules/products/Models/products.model.ts
export interface Products {
    id: number;
    product_type_id: number;
    name: string;
    description: string | null;
    price: string | null;
    stock: number | null;
    is_active: boolean | null;
    created_at: Date;
}

export interface ProductType {
    id: number;
    name: string;
}

export interface CreateProductDTO {
    product_type_id: number;
    name: string;
    description?: string;
    price?: string;
    stock?: number;
}

export interface UpdateProductDTO {
    product_type_id?: number;
    name?: string;
    description?: string;
    price?: string;
    stock?: number;
    is_active?: boolean;
}